+++ kit/dapp/src/components/asset-designer/asset-designer-form.tsx
@@ -1,12 +1,25 @@
 import { AssetBasics } from "@/components/asset-designer/asset-basics";
 import { SelectAssetType } from "@/components/asset-designer/select-asset-type";
-import { assetDesignerFormOptions } from "@/components/asset-designer/shared-form";
+import {
+  assetDesignerFormOptions,
+  AssetDesignerFormSchema,
+} from "@/components/asset-designer/shared-form";
 import { useAppForm } from "@/hooks/use-app-form";
 import { useStore } from "@tanstack/react-store";
 
 export const AssetDesignerForm = () => {
   const form = useAppForm({
     ...assetDesignerFormOptions,
+    validators: {
+      onChange: (values) => {
+        try {
+          const parseResult = AssetDesignerFormSchema.safeParse(values);
+          return parseResult.success;
+        } catch {
+          return false;
+        }
+      },
+    },
   });
   const step = useStore(form.store, (state) => state.values.step);
