--- kit/dapp/src/components/asset-designer/steps-schema.ts
@@ -1,14 +1,0 @@
-import { z } from "zod";
-
-export const AssetDesignerSteps = [
-  "selectAssetType",
-  "assetBasics",
-  "summary",
-] as const;
-
-const AssetDesignerStepsSchema = z.enum(AssetDesignerSteps);
-export type AssetDesignerSteps = z.infer<typeof AssetDesignerStepsSchema>;
-
-export const AssetDesignerStepSchema = z.object({
-  step: AssetDesignerStepsSchema,
-});
