+++ kit/dapp/src/components/asset-designer/steps.ts
@@ -1,4 +1,17 @@
-import type { AssetDesignerSteps } from "@/components/asset-designer/steps-schema";
+import { z } from "zod";
+
+export const AssetDesignerSteps = [
+  "selectAssetType",
+  "assetBasics",
+  "summary",
+] as const;
+
+const AssetDesignerStepsSchema = z.enum(AssetDesignerSteps);
+export type AssetDesignerSteps = z.infer<typeof AssetDesignerStepsSchema>;
+
+export const AssetDesignerStepSchema = z.object({
+  step: AssetDesignerStepsSchema,
+});
 
 export const assetDesignerSteps: Record<
   AssetDesignerSteps,
